audio: implement audio loopback on 4 endpoints
Implementation
--------------

USB is in high speed mode and declare via descriptors a USB audio class v1.0.

On Windows, this requires setting bInterval to 1ms (value = 4). It won't
work to transfer audio packets every micro-frame (125 µs).

The audio configuration is:

 - 4 audio out endpoint (using ISO OUT USB tokens).
 - 4 audio in endpoint that send to the host the same audio received on
   out endpoint (using ISO IN USB tokens).
 - All audio endpoints are 16 bits, 48000 Hz, Stereo, synchronous (no
   feedback endpoint).
 - All audio endpoints must have Mute + Volume controls, else Windows
   won't show detect them.

ISO transfer management
-------

Also, the STM32 require precise timing for USBD_LL_Transmit /
USBD_LL_PrepareReceive calls. They must occurs exactly between the
previous SOF and the SOF of the frame where the transfer must occur.

If the USB host doesn't issue a ISO transfer IN/OUT token, USBD_AUDIO_IsoINIncomplete
or USBD_AUDIO_IsoOutIncomplete will be called and the transfer must be re-prepared.
This can't be bypassed (by commenting code) else the USB hardware IP
seems stuck.

So this requires counting SOF frames to check whether we need to prepare
a new transfer for the next frame.

To detect the first ISO IN/OUT token issued by the USB host when
starting audio (which can happen at any frame, we can't predict it), the
code use the USBD_AUDIO_OutTokenWhileDisabled callback which is called
when receiving a USB token without having prepared a transfer (OTEPDM
interrupt flag).

This way, we know that at the frame 1ms later, there will be a ISO token
again (as audio is transfered every 1ms, as configured by bInterval in
USB descriptors).

Fixing ISO Incomplete events causing other ISO endpoint to be disabled
----------------------------------------------------------------------

There was an issue in HAL PCD driver about the handling of ISO
incomplete interrupt.
It was disabling too many ISO endpoint leading to glitches on running
audio endpoints when starting/stopping another endpoint.
The issue was probably mainly that we didn't set CGONAK when no endpoint
were aborted.
Also, we now also check for the frame number for ISO IN endpoints and
fixed the check of the frame number for ISO OUT endpoint (RegVal anded
value is not comparable to FrameNumber anded value directly).

Debug LED
---------

LED USER2 is ON when not in the USB interrupt.
So if it is near OFF, then we are almost all the time in the USB
interrupt (the CPU might not be able to keep up).

Issues left
-----------

The audio with Windows works fine but ISO incomplete events can still
occur while streaming audio for a long time (even if not starting /
stopping any audio). But there is no SOF count loss.

To debug this, STM32CubeIDE live watches can be used with debug
variables in USBD_AUDIO_SpeakerMicTypeDef.