damc: fix rare glitches on usb buffers
There was a cache invalidate done in USB interrupt while the SAI interrupt was writing to usb circular buffer.
This caused memory corruption as the data not yet commited to memory was thrown away by the cache invalidate.
The fix is to not use cache management functions when not using the buffer with the DMA.