usb: reduce IRQ flood from CDC
CDC use bulk endpoint which cause many NAK and OTEPD IRQ as the host will do busy polling the endpoint to send data.

The OutTokenWhileDisabled interrupt is needed for audio isochronous endpoint to find the correct frame number where the isochronous transfer must be done.

But it is not needed when audio is running correctly, so if it happens on a CDC endpoint, disable it.

It will be reenabled again by usbd_audio when an ISO incomplete IRQ is triggered.